+++ src/controllers/reviewsControllers/getReviewsInRoute.js
@@ -1,0 +1,33 @@
+const getReviewsInRoute = async (req, res) => {
+  const { routeId } = req.params
+
+  try {
+    const foundRoute = await route.getSingleRoute(routeId)
+
+    const allReviewsInRoute = await review.getAllReviewsFromRoute(
+      foundRoute._id
+    )
+
+    res.json({
+      message: 'success',
+      payload: {
+        data: allReviewsInRoute,
+        description: 'Reviews found successfully',
+        statusCode: 200,
+      },
+    })
+  } catch (err) {
+    console.log(err)
+
+    res.json({
+      message: 'failure',
+      error: {
+        err,
+        description: 'Could not get reviews.',
+        statusCode: 404,
+      },
+    })
+  }
+}
+
+export { getReviewsInRoute }
